replace depricated tslint with eslint and configurations and set up

+++ src/routes/expense-routes.ts
@@ -1,5 +1,6 @@
-import { IExpenseService, StatusCodes } from '../typings';
+import { IExpenseService } from '../typings';
 import { Request, Response } from 'express';
+import { check, validationResult } from 'express-validator';
 
 export class ExpenseRoutes {
   public static TAX = 15;
@@ -11,14 +12,26 @@ export class ExpenseRoutes {
   }
 
   initialize(server: any): void {
-    this.router.post('/expenses/', (req: Request, res: Response) =>
-      this.createExpense(req, res)
+    this.router.post(
+      '/expenses/',
+      [
+        check('description', 'Description is required').not().isEmpty(),
+        check('amount', 'Amount is Required').not().isEmpty(),
+        check('date', 'Date is Required').not().isEmpty()
+      ],
+      (req: Request, res: Response) => this.createExpense(req, res)
     );
     this.router.get('/expenses/', (req: Request, res: Response) =>
       this.getExpensies(req, res)
     );
-    this.router.patch('/expenses/:expenseId', (req: Request, res: Response) =>
-      this.updateExpense(req, res)
+    this.router.patch(
+      '/expenses/:expenseId',
+      [
+        check('description', 'Description is required').not().isEmpty(),
+        check('amount', 'Amount is Required').not().isEmpty(),
+        check('date', 'Date is Required').not().isEmpty()
+      ],
+      (req: Request, res: Response) => this.updateExpense(req, res)
     );
     this.router.delete('/expenses/:expenseId', (req: Request, res: Response) =>
       this.deleteExpense(req, res)
@@ -28,6 +41,12 @@ export class ExpenseRoutes {
 
   async createExpense(req: Request, res: Response) {
     try {
+      const errors = validationResult(req);
+      if (!errors.isEmpty()) {
+        return res.status(400).send({
+          errors: errors.array()
+        });
+      }
       const { body } = req;
       const newExpense = await this.expenseService.createExpense(body);
       res.status(201).send({
@@ -55,6 +74,12 @@ export class ExpenseRoutes {
 
   async updateExpense(req: Request, res: Response) {
     try {
+      const errors = validationResult(req);
+      if (!errors.isEmpty()) {
+        return res.status(400).send({
+          errors: errors.array()
+        });
+      }
       const { body } = req;
       const updatedExpense = await this.expenseService.updateExpense(body);
       res.status(200).send({
@@ -71,6 +96,11 @@ export class ExpenseRoutes {
   async deleteExpense(req: Request, res: Response) {
     try {
       const { expenseId } = req.params;
+      if (!expenseId) {
+        return res.send(400).send({
+          message: 'No Expense ID provided'
+        });
+      }
       const deletedExpense = await this.expenseService.deleteExpense(expenseId);
       res.status(200).json({
         expense: deletedExpense,
